+++ platform/macos/src/MGLMapView.h
@@ -21,8 +21,10 @@ typedef NS_OPTIONS(NSUInteger, MGLMapDebugMaskOptions) {
         diagnose collision and label placement issues. */
     MGLMapDebugCollisionBoxesMask = 1 << 4,
     
-    /** Overdraw inspector. */
-    MGLMapDebugOverdrawsMask = 1 << 5,
+    /** Each drawing operation is replaced by a translucent fill. Overlapping
+        drawing operations appear more prominent to help diagnose overdrawing.
+     */
+    MGLMapDebugOverdrawVisualizationMask = 1 << 5,
     
     /** The stencil buffer is shown instead of the color buffer. */
     MGLMapDebugStencilBufferMask = 1 << 6,